added some methods and linemedia model

+++ models/iblock/element.go
@@ -2,7 +2,7 @@ package iblock
 
 type Element struct {
 	Id       int64
-	MinPrice int64
+	MinPrice int32
 	Quantity int
 	Article  string
 	Brand    string
